ExtractRouteParamTypes moved to params file, updated ResolvedRoute so that when used without generic `params` is not `{}`
