+++ src/types/route.spec-d.ts
@@ -1,7 +1,8 @@
 import { describe, expectTypeOf, test } from 'vitest'
+import { ExtractRouteParamTypes } from '@/types/params'
 import { Path } from '@/types/path'
 import { Query } from '@/types/query'
-import { ExtractRouteParamTypes, Route } from '@/types/route'
+import { Route } from '@/types/route'
 
 describe('ExtractRouteParamTypes', () => {
   test('given routes with different params, some optional, combines into expected args for developer', () => {
